requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
boto==2.49.0
botocore==1.29.43
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
cycler==0.11.0
dj-database-url==0.5.0
Django==2.2
django-cors-headers==2.5.3
django-storages-redux==1.3.3
exceptiongroup==1.1.0
frozenlist==1.3.3
gensim==3.8.0
gunicorn==20.1.0
h11==0.14.0
idna==3.3
importlib-metadata==4.8.2
jmespath==1.0.1
joblib>=1.1.1
kiwisolver==1.4.4
matplotlib==3.7.2
multidict==6.0.4
nltk==3.6.5
numpy==1.25.2
openai==0.27.0
google-genai==0.6.0
outcome==1.2.0
packaging==23.0
pandas==2.0.3
patsy==0.5.3
Pillow==8.4.0
psycopg2-binary==2.9.5
pycparser==2.21
pyparsing==3.0.6
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
regex==2021.11.10
requests==2.26.0
s3transfer==0.6.0
scikit-learn==1.3.0
scikit-posthocs==0.7.0
scipy==1.11.1
seaborn==0.12.2
selenium==4.8.2
six==1.16.0
sklearn==0.0
smart-open==5.2.1
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
statsmodels==0.13.5
threadpoolctl==3.0.0
tqdm==4.62.3
trio==0.22.0
trio-websocket==0.9.2
trueskill==0.4.5
typing-extensions==4.0.0
urllib3==1.26.7
wsproto==1.2.0
yarl==1.8.2
yellowbrick==1.5
zipp==3.6.0
